requirements.txt
streamlit>=1.25.0
transformers>=4.35.0
torch>=2.0.0
soundfile>=0.12.1
numpy>=1.24.0
python-dotenv>=1.0.0
datasets>=2.14.0
speechbrain>=0.5.0
accelerate>=0.20.0
scipy>=1.10.0
librosa>=0.10.0
huggingface_hub>=0.19.0